+++ R/pkg_scaledic.R
@@ -27,7 +27,7 @@
 #' library(dplyr)
 #' dat %>%
 #'   select_scale(scale = "ITRF", subscale = "Ext") %>%
-#'   names2item(prefix = "subscale2", id = TRUE) %>%
+#'   names2item(prefix = c("subscale2", "index")) %>%
 #'   summarise_all(mean, na.rm = TRUE) %>%
 #'   round(2) %>%
 #'   t()