+++ R/select_scale.R
@@ -6,5 +6,5 @@
 #' @export
 select_scale <- function(data, scale = NULL, subscale = NULL, subscale_2 = NULL) {
   id <- get_index(data = data, scale = scale, subscale = subscale, subscale_2 = subscale_2)
-  data[, unique(id), drop = FALSE]
+  data[, id, drop = FALSE]
 }